fixed issue with forbidden show on tvdb
On Tvdb some show names are 'forbidden' and for those shows Tvdb returns
a name like ***403: Forbidden show **.
Autosub will crashed on that name if it is used on addic7ed. This fixed
now.
Added more logic for FINDING shownames ending with (yyyy), (UK), (US)
AND (BBC).

Added a lot of id lookups for addic7ed.
Added releasegroup 'Queens' and 'Juggalotus'
added hd_tv as equivalent for HDTV to support Sickrage renaming.
Change the showname lookup sequence so the namemapping is used first.